Fixed bugs from bootstrap
Bootstrap wasn't entirely removed from all files so was causing the rails server to operate abnormally. Bootstrap is now fully removed, and rails server operates as normal. Added some more html to most of the pages to make css styling easier. As well as creating a html template for the payments page. Although, I haven't created a way for the user to navigate to that page yet